refactor: split RULES registry into rules.ts
